Finish to extract all properties from demos components file
