+++ src/demos/components/components_properties.js
@@ -137,6 +137,180 @@ const GridTileProperties = [
   ["titleStyle", "object", "Override the inline-styles of the title element."],
 ];
 
+const LinearProperties = [
+  ["color", "string", "The color of the progress bar,  defaults to primary color of theme."],
+  ["max", "number", "The max value of progress,  only works in determinate mode."],
+  ["min", "number", "The min value of progress,  only works in determinate mode."],
+  ["mode", "enum:'determinate', 'indeterminate', 'indeterminate'", "The mode of show your progress,  indeterminate for when there is no value for progress."],
+  ["style", "object", "Override the inline-styles of the root element."],
+  ["value", "number", "The value of progress, only works in determinate mode."],
+];
+
+const ListProperties = [
+  ["children", "node", "These are usually ListItems that are passed to be part of the list."],
+  ["style", "object", "Override the inline-styles of the root element."],
+];
+
+const ListItemProperties = [
+  ["autoGenerateNestedIndicator", "bool", "If true,  generate a nested-list-indicator icon when nested list items are detected. Note that an indicator will not be created if a rightIcon or rightIconButton has been provided to the element."],
+  ["children", "node", "Children passed into the ListItem."],
+  ["containerElement", "union:stringelement'span'", "The element to use as the container for the ListItem. Either a string to use a DOM element or a ReactElement. This is useful for wrapping the ListItem in a custom Link component. If a ReactElement is given,  ensure that it passes all of its given props through to the underlying DOM element and renders its children prop for proper integration."],
+  ["disableKeyboardFocus", "bool", "If true,  the element will not be able to be focused by the keyboard."],
+  ["disabled", "bool", "If true,  the element will not be clickable and will not display hover effects. This is automatically disabled if either leftCheckbox or rightToggle is set."],
+  ["hoverColor", "string", "Override the hover background color."],
+  ["initiallyOpen", "bool", "If true,  the nested ListItems are initially displayed."],
+  ["innerDivStyle", "object", "Override the inline-styles of the inner div element."],
+  ["insetChildren", "bool", "If true,  the children will be indented by 72px. This is useful if there is no left avatar or left icon."],
+  ["isKeyboardFocused", "bool", "Use to control if the list item should render as keyboard focused. If undefined (default),  this will be automatically managed. If provided,  it will change the components style. Note that this will not change the actual focus - and should only be used when you want to simulate keyboard focus (eg. in a rich text input autocomplete)."],
+  ["leftAvatar", "element", "This is the Avatar element to be displayed on the left side."],
+  ["leftCheckbox", "element", "This is the Checkbox element to be displayed on the left side."],
+  ["leftIcon", "element", "This is the SvgIcon or FontIcon to be displayed on the left side."],
+  ["nestedItems", "arrayOf", "An array of ListItems to nest underneath the current ListItem."],
+  ["nestedLevel", "number", "Controls how deep a ListItem appears. This property is automatically managed,  so modify at your own risk."],
+  ["nestedListStyle", "object", "Override the inline-styles of the nested items' NestedList."],
+  ["onClick", "function", "Callback function fired when the list item is clicked.Signature: function(event: object) => void event: Click event targeting the list item."],
+  ["onKeyboardFocus", "function", "Callback function fired when the ListItem is focused or blurred by the keyboard. Signature: function(event: object,  isKeyboardFocused: boolean) => void event: focus or blur event targeting the ListItem. isKeyboardFocused: If true,  the ListItem is focused."],
+  ["onNestedListToggle", "function", "Callback function fired when the ListItem toggles its nested list.Signature: function(listItem: object) => voidlistItem: The ListItem."],
+  ["open", "bool", "Control toggle state of nested list."],
+  ["primaryText", "node", "This is the block element that contains the primary text. If a string is passed in,  a div tag will be rendered."],
+  ["primaryTogglesNestedList", "bool", "If true,  clicking or tapping the primary text of the ListItem toggles the nested list."],
+  ["rightAvatar", "element", "This is the Avatar element to be displayed on the right side."],
+  ["rightIcon", "element", "This is the SvgIcon or FontIcon to be displayed on the right side."],
+  ["rightIconButton", "element", "This is the IconButton to be displayed on the right side. Hovering over this button will remove the ListItem hover. Also,  clicking on this button will not trigger a ripple on the ListItem; the event will be stopped and prevented from bubbling up to cause a ListItem click."],
+  ["rightToggle", "element", "This is the Toggle element to display on the right side."],
+  ["secondaryText", "node", "This is the block element that contains the secondary text. If a string is passed in,  a div tag will be rendered."],
+  ["secondaryTextLines", "enum:1, 2", "Can be 1 or 2. This is the number of secondary text lines before ellipsis will show."],
+  ["style", "object", "Override the inline-styles of the root element."],
+];
+
+const RadioButtonProperties = [
+  ["checkedIcon", "element", "The icon element to show when the radio button is checked."],
+  ["disabled", "bool", "If true,  the radio button is disabled."],
+  ["iconStyle", "object", "Override the inline-styles of the icon element."],
+  ["inputStyle", "object", "Override the inline-styles of the input element."],
+  ["labelStyle", "object", "Override the inline-styles of the label element."],
+  ["style", "object", "Override the inline-styles of the root element."],
+  ["uncheckedIcon", "element", "The icon element to show when the radio button is unchecked."],
+  ["value", "any", "The value of the radio button."],
+];
+
+const RadioButtonGroupProperties = [
+  ["children", "node", "Should be used to pass RadioButton components."],
+  ["className", "string", "The CSS class name of the root element."],
+  ["defaultSelected", "any", "The value property of the radio button that will be selected by default. This takes precedence over the checked property of the RadioButton elements."],
+  ["labelPosition", "enum:'left', 'right'", "Where the label will be placed for all child radio buttons. This takes precedence over the labelPosition property of the RadioButton elements."],
+  ["name", "string", "The name that will be applied to all child radio buttons."],
+  ["onChange", "function", "Callback function that is fired when a radio button has been checked.Signature: function(event: object,  value: undefined) => void event: change event targeting the selected radio button. value: The value of the selected radio button."],
+  ["style", "object", "Override the inline-styles of the root element."],
+  ["valueSelected", "any", "The value of the currently selected radio button."],
+];
+
+const SelectfieldProperties = [
+  ["autoWidth", "bool", "If true,  the width will automatically be set according to the items inside the menu. To control the width in CSS instead,  leave this prop set to false."],
+  ["children", "node", "The MenuItem elements to populate the select field with. If the menu items have a label prop,  that value will represent the selected menu item in the rendered select field."],
+  ["disabled", "bool", "If true,  the select field will be disabled."],
+  ["dropDownMenuProps", "object", "Object that can handle and override any property of component DropDownMenu."],
+  ["errorStyle", "object", "Override the inline-styles of the error element."],
+  ["errorText", "node", "The error content to display."],
+  ["floatingLabelFixed", "bool", "If true,  the floating label will float even when no value is selected."],
+  ["floatingLabelStyle", "object", "Override the inline-styles of the floating label."],
+  ["floatingLabelText", "node", "The content of the floating label."],
+  ["fullWidth", "bool", "If true,  the select field will take up the full width of its container."],
+  ["hintStyle", "object", "Override the inline-styles of the hint element."],
+  ["hintText", "node", "The hint content to display."],
+  ["iconStyle", "object", "Override the inline-styles of the icon element."],
+  ["id", "string", "The id prop for the text field."],
+  ["labelStyle", "object", "Override the label style when the select field is inactive."],
+  ["listStyle", "object", "Override the inline-styles of the underlying List element."],
+  ["maxHeight", "number", "Override the default max-height of the underlying DropDownMenu element."],
+  ["menuItemStyle", "object", "Override the inline-styles of menu items."],
+  ["menuStyle", "object", "Override the inline-styles of the underlying DropDownMenu element."],
+  ["multiple", "bool", "If true,  value must be an array and the menu will support multiple selections."],
+  ["onChange", "function", "Callback function fired when a menu item is selected. Signature: function(event: object,  key: number,  payload: any) => void event: Click event targeting the menu item that was selected. key: The index of the selected menu item,  or undefined if multiple is true. payload: If multiple is true,  the menu's value array with either the menu item's value added (if it wasn't already selected) or omitted (if it was already selected). Otherwise,  the value of the menu item."],
+  ["selectedMenuItemStyle", "object", "Override the inline-styles of selected menu items."],
+  ["selectionRenderer", "function", "Customize the rendering of the selected item. Signature: function(value: any,  menuItem: any) => void value: If multiple is true,  the menu's value array with either the menu item's value added (if it wasn't already selected) or omitted (if it was already selected). Otherwise,  the value of the menu item. menuItem: The selected MenuItem. If multiple is true,  this will be an array with the MenuItems matching the values parameter."],
+  ["style", "object", "Override the inline-styles of the root element."],
+  ["underlineDisabledStyle", "object", "Override the inline-styles of the underline element when the select field is disabled."],
+  ["underlineFocusStyle", "object", "Override the inline-styles of the underline element when the select field is focused."],
+  ["underlineStyle", "object", "Override the inline-styles of the underline element."],
+  ["value", "any", "If multiple is true, an array of the values of the selected menu items. Otherwise, the value of the selected menu item. If provided, the menu will be a controlled component."],
+];
+
+const SliderProperties = [
+  ["axis", "enum: 'x',  'x-reverse',  'y',  'y-reverse',  'x'", "The axis on which the slider will slide."],
+  ["defaultValue", "valueInRangePropType", "The default value of the slider."],
+  ["disableFocusRipple", "bool", "Disables focus ripple if set to true."],
+  ["disabled", "bool", "If true,  the slider will not be interactable."],
+  ["max", "minMaxPropType", "The maximum value the slider can slide to on a scale from 0 to 1 inclusive. Cannot be equal to min."],
+  ["min", "minMaxPropType", "The minimum value the slider can slide to on a scale from 0 to 1 inclusive. Cannot be equal to max."],
+  ["name", "string", "The name of the slider. Behaves like the name attribute of an input element."],
+  ["onChange", "function", "Callback function that is fired when the slider's value changed. Signature: function(event: object,  newValue: number) => void event: KeyDown event targeting the slider. newValue: The new value of the slider."],
+  ["onDragStart", "function", "Callback function that is fired when the slider has begun to move. Signature: function(event: object) => void event: MouseDown or TouchStart event targeting the slider."],
+  ["onDragStop", "function", "Callback function that is fired when the slide has stopped moving. Signature: function(event: object) => void event: MouseEnd or TouchEnd event targeting the slider."],
+  ["required", "bool", "Whether or not the slider is required in a form."],
+  ["sliderStyle", "object", "Override the inline-styles of the inner slider element."],
+  ["step", "number", "The granularity the slider can step through values."],
+  ["style", "object", "Override the inline-styles of the root element."],
+  ["value", "valueInRangePropType", "The value of the slider."],
+];
+
+const TabsProperties = [
+  ["children", "node", "Should be used to pass Tab components."],
+  ["className", "string", "The css class name of the root element."],
+  ["contentContainerClassName", "string", "The css class name of the content's container."],
+  ["contentContainerStyle", "object", "Override the inline-styles of the content's container."],
+  ["initialSelectedIndex", "number", "Specify initial visible tab index. If initialSelectedIndex is set but larger than the total amount of specified tabs,  initialSelectedIndex will revert back to default. If initialSelectedIndex is set to any negative value,  no tab will be selected intially."],
+  ["inkBarStyle", "object", "Override the inline-styles of the InkBar."],
+  ["onChange", "function", "Called when the selected value change."],
+  ["style", "object", "Override the inline-styles of the root element."],
+  ["tabItemContainerStyle", "object", "Override the inline-styles of the tab-labels container."],
+  ["tabTemplate", "function", "Override the default tab template used to wrap the content of each tab element."],
+  ["tabTemplateStyle", "object", "Override the inline-styles of the tab template."],
+  ["value", "any", "Makes Tabs controllable and selects the tab whose value prop matches this prop."],
+];
+
+const TabProperties = [
+  ["buttonStyle", "object", "Override the inline-styles of the button element."],
+  ["className", "string", "The css class name of the root element."],
+  ["icon", "node", "Sets the icon of the tab,  you can pass FontIcon or SvgIcon elements."],
+  ["label", "node", "Sets the text value of the tab item to the string specified."],
+  ["onActive", "function", "Fired when the active tab changes by touch or tap. Use this event to specify any functionality when an active tab changes. For example - we are using this to route to home when the third tab becomes active. This function will always recieve the active tab as it's first argument."],
+  ["style", "object", "Override the inline-styles of the root element."],
+  ["value", "any", "If value prop passed to Tabs component, this value prop is also required. It assigns a value to the tab so that it can be selected by the Tabs."],
+];
+
+const TextfieldProperties = [
+  ["children", "node", "The textfield extended by this."],
+  ["className", "string", "The css class name of the root element."],
+  ["defaultValue", "any", "The text string to use for the default value."],
+  ["disabled", "bool", "Disables the text field if set to true."],
+  ["errorStyle", "object", "The style object to use to override error styles."],
+  ["errorText", "node", "The error content to display."],
+  ["floatingLabelFixed", "bool", "If true,  the floating label will float even when there is no value."],
+  ["floatingLabelFocusStyle", "object", "The style object to use to override floating label styles when focused."],
+  ["floatingLabelShrinkStyle", "object", "The style object to use to override floating label styles when shrunk."],
+  ["floatingLabelStyle", "object", "The style object to use to override floating label styles."],
+  ["floatingLabelText", "node", "The content to use for the floating label element."],
+  ["fullWidth", "bool", "If true,  the field receives the property width 100%."],
+  ["hintStyle", "object", "Override the inline-styles of the TextField's hint text element."],
+  ["hintText", "node", "The hint content to display."],
+  ["id", "string", "The id prop for the text field."],
+  ["inputStyle", "object", "Override the inline-styles of the TextField's input element. When multiLine is false: define the style of the input element. When multiLine is true: define the style of the container of the textarea."],
+  ["multiLine", "bool", "If true,  a textarea element will be rendered. The textarea also grows and shrinks according to the number of lines."],
+  ["name", "string", "Name applied to the input."],
+  ["onChange", "function", "Callback function that is fired when the textfield's value changes. Signature: function(event: object,  newValue: string) => void event: Change event targeting the text field. newValue: The new value of the text field."],
+  ["rows", "number", "Number of rows to display when multiLine option is set to true."],
+  ["rowsMax", "number", "Maximum number of rows to display when multiLine option is set to true."],
+  ["style", "object", "Override the inline-styles of the root element."],
+  ["textareaStyle", "object", "Override the inline-styles of the TextField's textarea element. The TextField use either a textarea or an input,  this property has effects only when multiLine is true."],
+  ["type", "string", "Specifies the type of input to display such as 'password' or 'text'."],
+  ["underlineDisabledStyle", "object", "Override the inline-styles of the TextField's underline element when disabled."],
+  ["underlineFocusStyle", "object", "Override the inline-styles of the TextField's underline element when focussed."],
+  ["underlineShow", "bool", "If true,  shows the underline for the text field."],
+  ["underlineStyle", "object", "Override the inline-styles of the TextField's underline element."],
+  ["value", "any", "The value of the text field."],
+];
+
 export {
   head,
   ButtonsProperties,
@@ -149,4 +323,14 @@ export {
   CheckboxProperties,
   GridListProperties,
   GridTileProperties,
+  LinearProperties,
+  ListProperties,
+  ListItemProperties,
+  RadioButtonProperties,
+  RadioButtonGroupProperties,
+  SelectfieldProperties,
+  SliderProperties,
+  TabsProperties,
+  TabProperties,
+  TextfieldProperties,
 };
