+++ src/demos/components/textfields.jsx
@@ -6,42 +6,7 @@
  */
 import React from "react";
 import { TextField, Grid, Inner, Cell } from "../../../src/libs";
-
-const head = [
-  "Name", "Type", "Desc",
-];
-
-const TextfieldProperties = [
-  ["children", "node", "The textfield extended by this."],
-  ["className", "string", "The css class name of the root element."],
-  ["defaultValue", "any", "The text string to use for the default value."],
-  ["disabled", "bool", "Disables the text field if set to true."],
-  ["errorStyle", "object", "The style object to use to override error styles."],
-  ["errorText", "node", "The error content to display."],
-  ["floatingLabelFixed", "bool", "If true,  the floating label will float even when there is no value."],
-  ["floatingLabelFocusStyle", "object", "The style object to use to override floating label styles when focused."],
-  ["floatingLabelShrinkStyle", "object", "The style object to use to override floating label styles when shrunk."],
-  ["floatingLabelStyle", "object", "The style object to use to override floating label styles."],
-  ["floatingLabelText", "node", "The content to use for the floating label element."],
-  ["fullWidth", "bool", "If true,  the field receives the property width 100%."],
-  ["hintStyle", "object", "Override the inline-styles of the TextField's hint text element."],
-  ["hintText", "node", "The hint content to display."],
-  ["id", "string", "The id prop for the text field."],
-  ["inputStyle", "object", "Override the inline-styles of the TextField's input element. When multiLine is false: define the style of the input element. When multiLine is true: define the style of the container of the textarea."],
-  ["multiLine", "bool", "If true,  a textarea element will be rendered. The textarea also grows and shrinks according to the number of lines."],
-  ["name", "string", "Name applied to the input."],
-  ["onChange", "function", "Callback function that is fired when the textfield's value changes. Signature: function(event: object,  newValue: string) => void event: Change event targeting the text field. newValue: The new value of the text field."],
-  ["rows", "number", "Number of rows to display when multiLine option is set to true."],
-  ["rowsMax", "number", "Maximum number of rows to display when multiLine option is set to true."],
-  ["style", "object", "Override the inline-styles of the root element."],
-  ["textareaStyle", "object", "Override the inline-styles of the TextField's textarea element. The TextField use either a textarea or an input,  this property has effects only when multiLine is true."],
-  ["type", "string", "Specifies the type of input to display such as 'password' or 'text'."],
-  ["underlineDisabledStyle", "object", "Override the inline-styles of the TextField's underline element when disabled."],
-  ["underlineFocusStyle", "object", "Override the inline-styles of the TextField's underline element when focussed."],
-  ["underlineShow", "bool", "If true,  shows the underline for the text field."],
-  ["underlineStyle", "object", "Override the inline-styles of the TextField's underline element."],
-  ["value", "any", "The value of the text field."],
-];
+import { head, TextfieldProperties } from "./components_properties";
 
 export default () => (
   <section>
